fixup based upon review comments
- moved around the code so that the helpers are accessible from the lib
  dir
- created beaker-rspec.rb that pulls in all necessary files to run tests
