+++ sample.cfg
@@ -6,18 +6,18 @@ HOSTS:
       - dashboard
       - database
     platform: ubuntu-10.04-amd64
-    hypervisor : vagrant
+    hypervisor: vagrant
     box: ubuntu-server-10044-x64-vbox4210
-    box_url : http://puppet-vagrant-boxes.puppetlabs.com/ubuntu-server-10044-x64-vbox4210.box
-    ip : 192.168.20.20
+    box_url: http://puppet-vagrant-boxes.puppetlabs.com/ubuntu-server-10044-x64-vbox4210.box
+    ip: 192.168.20.20
   ubuntu-10-04-4-x64-agent:
     roles:
       - agent
     platform: ubuntu-10.04-amd64
-    hypervisor : vagrant
+    hypervisor: vagrant
     box: ubuntu-server-10044-x64-vbox4210
-    box_url : http://puppet-vagrant-boxes.puppetlabs.com/ubuntu-server-10044-x64-vbox4210.box
-    ip : 192.168.21.21
+    box_url: http://puppet-vagrant-boxes.puppetlabs.com/ubuntu-server-10044-x64-vbox4210.box
+    ip: 192.168.21.21
 CONFIG:
   nfs_server: none
   consoleport: 443
