--- spec/helpers.rb
@@ -1,7 +1,0 @@
-module BeakerRSpec::Helpers
-  include Beaker::DSL
-
-  def hosts
-    self.class.hosts
-  end
-end
